+++ tcvectordb/rpc_base_index.go
@@ -23,3 +23,8 @@ func (r *rpcImplementerIndex) RebuildIndex(ctx context.Context, params ...*Rebui
 func (r *rpcImplementerIndex) AddIndex(ctx context.Context, params ...*AddIndexParams) error {
 	return r.flat.AddIndex(ctx, r.database.DatabaseName, r.collection.CollectionName, params...)
 }
+
+// [ModifyVectorIndex] modifies vector indexes to an existing collection.
+func (r *rpcImplementerIndex) ModifyVectorIndex(ctx context.Context, param ModifyVectorIndexParam) error {
+	return r.flat.ModifyVectorIndex(ctx, r.database.DatabaseName, r.collection.CollectionName, param)
+}